refactor: 4단계: HttpResponse 도입
